+++ src/app/station/station-detail.component.html
@@ -9,8 +9,8 @@
         <div class="station-detail__name">{{ (detailSubject$ | async).source | translate }}</div>
         <div class="station-detail__label">{{ 'Höhe' | translate }}</div>
         <div>
-          <span class="station-detail__value">{{ (detailSubject$ | async)?.altitude }}</span><span
-            class="station-detail__value-addition" translate> {{ 'm ü. M.' }}</span>
+          <span class="station-detail__value">{{ (detailSubject$ | async)?.altitude }}</span
+          ><span class="station-detail__value-addition" translate> {{ 'm ü. M.' }}</span>
         </div>
       </div>
       <div class="station-detail__observation-content">
@@ -30,13 +30,20 @@
             <ng-container *ngIf="item.observation.isSome()">
               <div class="station-detail__label">{{ item.phenophase.de | translate }}</div>
               <mat-form-field class="station-detail__form-field">
-                <input matInput [id]="item.phenophase.id" [matDatepicker]="fakePicker"
+                <input
+                  matInput
+                  [id]="item.phenophase.id"
+                  [matDatepicker]="fakePicker"
                   placeholder="{{ item.phenophase.description_de | translate }}"
-                  [value]="item.observation.toNullable()?.date" readonly="true" />
+                  [value]="item.observation.toNullable()?.date"
+                  readonly="true"
+                />
               </mat-form-field>
               <mat-datepicker #fakePicker></mat-datepicker>
-              <div class="statiohn-detail__detail-sub"
-                *ngIf="item.observation.toNullable()?.comment?.toString() as comment">
+              <div
+                class="statiohn-detail__detail-sub"
+                *ngIf="item.observation.toNullable()?.comment?.toString() as comment"
+              >
                 {{ 'Kommentar' | translate }}: {{ staticComments[comment] | translate }}
               </div>
             </ng-container>
